Adding multiple views/outputs per elm file
This PR adds support for grouping multiple views/output files.

This is useful when many of your views share depenencies: these will now
be compiled only once. This is a double-edged sword because when you
want to regenerate one of the views all the code has to be compiled.

Added is the function `elmStaticHtmlMultiple`
```typescript
elmStaticHtmlMultiple(
    rootDir: string,
    moduleName: string,
    configs: ViewFunctionConfig[],
    alreadyRun?: boolean,
    elmMakePath?: string,
    installMethod?: string
    ): Promise<void[]>
```

calling works like

```typescript
const configs =
    [ { viewFunction: "MyModule.view", model: model
        , decoder: "MyModule.decodeModel", output: "multiple1.html" }
        , { viewFunction: "MyModule.lazyView", model: model
        , decoder: "MyModule.decodeModel", output: "multiple2.html" }
    ];

elmStaticHtmlMultiple(process.cwd(), "MyModule", configs);
```

A major difference with `elmStaticHtml` is that the output file is part
of the config and consequently a `Promise<void>` is returned.

The module name (2nd argument) is only there to create a unique hash

I'm sure there is a bunch of style things that need to be changed, but
does the basic architecture make sense?