also hash the produced elm.js file
This seems to be needed because of asynchronicity. Removing it
will cause the `elm.js` file to be overwritten (by the next view/model
pair) before it can be run to produce output.

Normally we might be able to compile multiple apps into one elm.js file,
but the port namespace still seems to be global and port names are
linked to view names, so if the same view function is used multiple
times, this will cause an error. (clash in port names).