allow grouping multiple view functions into one elm file (#8)
* Adding multiple views/outputs per elm file

* rename to grouped

* Make grouped return a Promise<[string]>

Also removing the `output` field from the config for grouped.

* Only compile files required for a step

This fixes a problem of clashing port names when a view function
is used multiple times in a grouped and a normal call of elmStaticHtml

* also hash the produced elm.js file

This seems to be needed because of asynchronicity. Removing it
will cause the `elm.js` file to be overwritten (by the next view/model
pair) before it can be run to produce output.

Normally we might be able to compile multiple apps into one elm.js file,
but the port namespace still seems to be global and port names are
linked to view names, so if the same view function is used multiple
times, this will cause an error. (clash in port names).

* make default use the grouped code under the hood

* rename grouped to multiple

* thread the filename through

* return a { generatedHtml, fileOutputName } record

Friendlier than a 2-element array. Also fixes a bug where
the model was not defined in a config.

* switch to fileOutputName everywhere

* add to README